refactor(rover_py): move data records to __init__.py
- Put commonly used data records in __init__.py so they are availabe using `import rover`
- Split out bootloader-specifics into bootloader.py
- Fix all files affected by changes
